refactor: refer to variables with copyable IDs
Added methods of finding variables. Each owned Value is stored in an
arena in its owning scope, and referred to by a `ValueId`. They are referenced in the owning
scope's variable arena with `Var::Local(ValueId)`. If a scope needs to reference a variable
outside of its scope, it will store a `Var::Reference(ScopeId, VarId)`. This
way, the `ScopeManager::find_var` and `ScopeManager::resolve_reference`
functions can be used to retrieve a reference to an actual value, even
if several references have to be followed to get to it.